Change net field type to BigDecimal in DaybookResponse
Updated the 'net' field in the Totals class to use BigDecimal for better precision.

+++ src/main/java/com/example/bullion/model/DaybookResponse.java
@@ -18,7 +18,8 @@ public class DaybookResponse {
     private int purchase;
     private int cashin;
     private int cashout;
-    private int net;
+    private BigDecimal net;
   }
 }
+
